Update devise to prompt for install
Also prompt at the start of the whole install procedure to tell the user that we’ll be running “bundle install” in their current context and give them an option to stop and get the right ruby and/or gemset if they need to.

Additionally, don’t use devise templates, use thor’s inject_into_file to prepend some file inputs (conditionally) to the newly generated devise slim views.

Also fix an issue where the spring binstubs being generated by suspenders (as their last thing) would stop generators from working for us. So don’t use their spring_setup and instead do all our stuff, then do our own version of their same spring_setup.

+++ lib/voyage/app_builder.rb
@@ -1,34 +1,51 @@
 module Suspenders
   class AppBuilder < Rails::AppBuilder
     def application_js
-      template "../templates/application.js", "app/assets/javascripts/application.js", force: true
+      template '../templates/application.js', 'app/assets/javascripts/application.js', force: true
     end
 
     def application_controller
-      template "../templates/application_controller.rb", "app/controllers/application_controller.rb", force: true
+      template '../templates/application_controller.rb', 'app/controllers/application_controller.rb', force: true
     end
 
     def install_devise
-      bundle_command 'exec rails generate devise:install'
-      bundle_command 'exec rails generate scaffold user first_name:string last_name:string'
-      bundle_command 'exec rails generate devise user'
-      bundle_command 'exec rails generate devise:views'
-      run 'gem install html2slim'
-      inside('lib') do # arbitrary, run in context of newly generated app
-        run "erb2slim '../app/views/devise' './app/views/devise'"
-        run "erb2slim -d '../app/views/devise'"
+      if yes?('Would you like to install Devise? (y/N)')
+        bundle_command 'exec rails generate devise:install'
+
+        model_name = ask('What would you like the user model to be called? [user]')
+        model_name = 'user' if model_name.blank?
+
+        if yes?("Would you like to add first_name and last_name to the devise model? (y/N)")
+          adding_first_and_last_name = true
+          bundle_command "exec rails generate scaffold #{model_name} first_name:string last_name:string"
+        end
+
+        bundle_command "exec rails generate devise #{model_name}"
+        bundle_command 'exec rails generate devise:views'
+
+        run 'gem install html2slim'
+        inside('lib') do # arbitrary, run in context of newly generated app
+          run "erb2slim '../app/views/devise' '../app/views/devise'"
+          run "erb2slim -d '../app/views/devise'"
+        end
+
+        customize_devise_views if adding_first_and_last_name
       end
     end
 
-    def custom_devise_views
-      template "../templates/devise/edit.html.slim", "app/views/devise/registrations/edit.html.slim", force: true
-      template "../templates/devise/new.html.slim", "app/views/devise/registrations/new.html.slim", force: true
+    def customize_devise_views
+      %w(edit new).each do |file|
+        file_path = "app/views/devise/registrations/#{file}.html.slim"
+        inject_into_file file_path, before: "    = f.input :email, required: true, autofocus: true" do <<-'RUBY'
+    = f.input :first_name, required: true, autofocus: true
+    = f.input :last_name, required: true
+RUBY
+        end
+      end
     end
 
-
-
     def gemfile
-      template "../templates/Gemfile.erb", "Gemfile"
+      template '../templates/Gemfile.erb', 'Gemfile'
     end
 
     def configure_generators
